# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для управления балансом пользователей и просмотра заказов.",
        "title": "API для управления бонусным балансом Гофермарта",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/user/withdrawals": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает JSON-массив всех списания текущего пользователя",
                "tags": [
                    "balance"
                ],
                "summary": "Получить список всех списаний",
                "responses": {
                    "200": {
                        "description": "Успешное списание",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Operation"
                            }
                        }
                    },
                    "204": {
                        "description": "Отсутствуют операции",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/balance": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает в виде JSON текущий баланс для текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Получить текущий баланс пользователя",
                "responses": {
                    "200": {
                        "description": "Успешное получение списка заказов",
                        "schema": {
                            "$ref": "#/definitions/model.Balance"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/balance/withdraw": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Вызов эндпойнта приводит к списанию баллов с бонусного баланса, если их достаточно",
                "tags": [
                    "balance"
                ],
                "summary": "Списать бонусные баллы",
                "parameters": [
                    {
                        "description": "Данные для списания (номер заказа и сумма)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Operation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное списание",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "Недостаточно бонусов нв балансе",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Balance": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "withdrawn": {
                    "type": "integer"
                }
            }
        },
        "model.Operation": {
            "type": "object",
            "properties": {
                "order": {
                    "type": "string"
                },
                "sum": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "auth_token",
            "in": "cookie"
        }
    }
}